Logic & Logging in NodeWebController and GetCurrentUserId UserService…

+++ api/SchemaStar/SchemaStar/Controllers/NodewebsController.cs
@@ -2,50 +2,66 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using SchemaStar.DTOs.Nodeweb_DTOs;
+using SchemaStar.Exceptions;
 using SchemaStar.Models;
+using SchemaStar.Services;
 
 namespace SchemaStar.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class NodewebsController : ControllerBase
     {
         private readonly SchemastarContext _context;
+        private readonly IUserService _userService;
+        private readonly ILogger<NodewebsController> _logger;
 
-        public NodewebsController(SchemastarContext context)
+        public NodewebsController(SchemastarContext context, IUserService userService, ILogger<NodewebsController> logger)
         {
             _context = context;
+            _userService = userService;
+            _logger = logger;
         }
 
         // GET: api/Nodewebs
         [HttpGet]
         public async Task<ActionResult<IEnumerable<NodewebResponseDTO>>> GetNodewebs()
         {
-            return await _context.Nodewebs.Select(n => new NodewebResponseDTO
-            { 
-                PublicId = new Guid (n.PublicId),
-                NodeWebName = n.NodeWebName,
-                CreatedAt = n.CreatedAt,
-                UpdatedAt = n.UpdatedAt,
-                LastLayoutAt = n.LastLayoutAt,
-            }).ToListAsync();
+            var userId = _userService.GetCurrentUserId();
+            if (userId == null) throw new UnauthorizedException("User does not have permission to access these NodeWebs");
+             
+            return await _context.Nodewebs
+                .Where(n => n.UserId == userId)
+                .Select(n => new NodewebResponseDTO
+                {
+                    PublicId = n.PublicId.ToGuidFromMySqlBinary(),
+                    NodeWebName = n.NodeWebName,
+                    CreatedAt = n.CreatedAt,
+                    UpdatedAt = n.UpdatedAt,
+                    LastLayoutAt = n.LastLayoutAt,
+                }).ToListAsync();
         }
 
         // GET: api/Nodewebs/{guid}
         [HttpGet("{publicId}")]
         public async Task<ActionResult<NodewebResponseDTO>> GetNodeweb(Guid publicId)
         {
-            byte[] publicIdBytes = publicId.ToByteArray();
+            var userId = _userService.GetCurrentUserId();
+            if (userId == null) throw new UnauthorizedException("User does not have permission to access this NodeWeb");
+
+            byte[] publicIdBytes = publicId.ToMySqlBinary();
 
             var nodeweb = await _context.Nodewebs
-                .Where(n => n.PublicId == publicIdBytes)
+                .Where(n => n.PublicId == publicIdBytes && n.UserId == userId)
                 .Select(n => new NodewebResponseDTO
                 {
-                    PublicId = new Guid(n.PublicId),
+                    PublicId = n.PublicId.ToGuidFromMySqlBinary(),
                     NodeWebName = n.NodeWebName,
                     CreatedAt = n.CreatedAt,
                     UpdatedAt = n.UpdatedAt,
@@ -54,23 +70,30 @@ namespace SchemaStar.Controllers
 
             if (nodeweb == null)
             {
-                return NotFound();
+                _logger.LogWarning("NodeWeb {PublicId} not found for User {UserId}", publicId, userId);
+                throw new NotFoundException("NodeWeb does exists for the user");
             }
-
+            
             return nodeweb;
         }
 
-        // PUT: api/Nodewebs/5
+        // PATCH: api/Nodewebs/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("{publicId}")]
+        [HttpPatch("{publicId}")]
         public async Task<IActionResult> UpdateNodeweb(Guid publicId, NodewebRequestDTO request)
         {
-            byte[] publicIdBytes = publicId.ToByteArray();
-            var nodeweb = await _context.Nodewebs.FirstOrDefaultAsync(n => n.PublicId == publicIdBytes);
-           
+            var userId = _userService.GetCurrentUserId();
+            if (userId == null) throw new UnauthorizedException("User does not have permission to modify this NodeWeb"); 
+
+            byte[] publicIdBytes = publicId.ToMySqlBinary();
+            var nodeweb = await _context.Nodewebs
+                .Where(n => n.UserId == userId)
+                .FirstOrDefaultAsync(n => n.PublicId == publicIdBytes);
+
             if (nodeweb == null)
             {
-                return NotFound();
+                _logger.LogWarning("NodeWeb {PublicId} for User {UserId} not found", publicId, userId);
+                throw new NotFoundException("NodeWeb does not exist for this user");
             }
 
             //Check for duplicate nodeweb names per the user
@@ -80,7 +103,11 @@ namespace SchemaStar.Controllers
              n.PublicId != publicIdBytes
             );
 
-            if (isDuplicate) { return BadRequest("Nodeweb name already exist, duplicate found."); }
+            if (isDuplicate)
+            {
+                _logger.LogWarning("Duplicate name for NodeWeb {PublicId} found for User {UserId}", nodeweb.PublicId, userId);
+                throw new ConflictException("NodeWeb with this name already exists");
+            }
 
             //Update the name
             nodeweb.NodeWebName = request.NodeWebName;
@@ -88,14 +115,17 @@ namespace SchemaStar.Controllers
             try
             {
                 await _context.SaveChangesAsync();
+                _logger.LogInformation("NodeWeb: {PublicId} has been updated by {UserId}", nodeweb.PublicId, userId);
             }
             catch (DbUpdateConcurrencyException)
             {
-                return Conflict("Record was updated elsewhere [database]");
+                _logger.LogError("Concurrency Error: Database update for NodeWeb {PublicId} conflicting with an existing update", nodeweb.PublicId);
+                throw new ConflictException("Record was updated elsewhere by another process");
             }
             catch (DbUpdateException) 
             {
-                return BadRequest("Unable to save, make sure node web name is unqiue.");
+                _logger.LogError("NodeWeb {PublicId} name is invalid make sure the name is unique", nodeweb.PublicId);
+                throw new ArgumentException("Unable to save, make sure node web name is unqiue.");
             }
 
             return NoContent();
@@ -106,46 +136,70 @@ namespace SchemaStar.Controllers
         [HttpPost]
         public async Task<ActionResult<NodewebResponseDTO>> PostNodeweb(NodewebRequestDTO request)
         {
-            ulong currentUserId = 0; //Implement logic for getting authenticated user
+            var userId = _userService.GetCurrentUserId();
+            if (userId == null) throw new UnauthorizedException("User does not have permission to create NodeWeb");
 
             //Check if nodeweb with the request name already created for the user
             bool isDuplicate = await _context.Nodewebs.
-                AnyAsync(n => n.UserId == currentUserId && n.NodeWebName == request.NodeWebName);
+                AnyAsync(n => n.UserId == userId && n.NodeWebName == request.NodeWebName);
 
-            if (isDuplicate) { return Conflict(new { message = $"A NodeWeb name '{request.NodeWebName} already exists for this user.'" }); }
+            if (isDuplicate)
+            {
+                _logger.LogWarning("Duplicate NodeWeb name found could not create new NodeWeb with name {NodeWebName} for User {UserId}", request.NodeWebName, userId);
+                throw new ConflictException("NodeWeb name already exists for this user");
+            }
+
+            var newGuid = Guid.NewGuid(); //Generate a new GUID and convert to MySQL binary format
 
             var nodeweb = new Nodeweb
             {
-                UserId = currentUserId,
+                PublicId = newGuid.ToMySqlBinary(),
                 NodeWebName = request.NodeWebName,
+                UserId = (ulong) userId
             };
 
+            var response = new NodewebResponseDTO
+            {
+                PublicId = newGuid,
+                NodeWebName = nodeweb.NodeWebName,
+                CreatedAt = nodeweb.CreatedAt,
+                UpdatedAt = nodeweb.UpdatedAt,
+            };
+            
             _context.Nodewebs.Add(nodeweb);
+            await _context.SaveChangesAsync();
 
-            return CreatedAtAction("GetNodeweb", new { id = nodeweb.Id }, nodeweb);
+            _logger.LogInformation("NodeWeb {PublicId} created for User {UserId}", response.PublicId, userId);
+
+            return CreatedAtAction(nameof(GetNodeweb), new { publicId = response.PublicId }, response);
         }
 
         // DELETE: api/Nodewebs/{guid}
         [HttpDelete("{publicId}")]
         public async Task<IActionResult> DeleteNodeweb(Guid publicId)
         {
-            byte[] publicIdBytes = publicId.ToByteArray();
+            var userId = _userService.GetCurrentUserId();
+            if (userId == null) throw new UnauthorizedException("User does not have permission to delete NodeWeb");
 
-            var nodeweb = await _context.Nodewebs.FirstOrDefaultAsync(n => n.PublicId == publicIdBytes);
+            byte[] publicIdBytes = publicId.ToMySqlBinary();
+
+            var nodeweb = await _context.Nodewebs
+                .Where(n => n.UserId == userId)
+                .FirstOrDefaultAsync(n => n.PublicId == publicIdBytes);
+
             if (nodeweb == null)
             {
-                return NotFound();
+                _logger.LogWarning("NodeWeb {PublicId} not found for User {UserId}", publicId, userId);
+                throw new NotFoundException("NodeWeb does not exists");
             }
-
+            
             _context.Nodewebs.Remove(nodeweb);
             await _context.SaveChangesAsync();
+
+            _logger.LogInformation("NodeWeb {PublicId} has been deleted by User {UserId}", nodeweb.PublicId, userId);
 
             return NoContent();
         }
 
-        private bool NodewebExists(ulong id)
-        {
-            return _context.Nodewebs.Any(e => e.Id == id);
-        }
     }
 }
